refactor(frontend): split MutualFundProjection, NetWorth, Budget pages and demo data under 500 LOC
Each oversized file split into a thin orchestrator + hook + sub-components,
matching the existing convention used by bill-calendar, goals, and tax-planning pages.

- MutualFundProjectionPage: 930 -> 110 LOC (orchestrator only)
- NetWorthPage: 884 -> 139 LOC
- BudgetPage: 860 -> 241 LOC
- generateDerivedData: 863 -> 27 LOC barrel
- generateTransactions: 692 -> 35 LOC orchestrator

Behavior preserved; type-check passes.
