+++ frontend/src/lib/demo/demoTemplates.ts
@@ -1,0 +1,105 @@
+export const ACCOUNTS = {
+  sbi: 'SBI Savings',
+  hdfc: 'HDFC Salary',
+  axis: 'Axis Bank',
+  swiggyCC: 'Swiggy HDFC Credit Card',
+  amazonCC: 'Amazon Pay ICICI Credit Card',
+  axisCC: 'Flipkart Axis Credit Card',
+  gpay: 'GPay UPI',
+  pluxee: 'Pluxee Wallet',
+  amazonWallet: 'Amazon Wallet',
+  growStocks: 'Groww Stocks',
+  growMF: 'Groww Mutual Funds',
+  epf: 'EPF Account',
+  ppf: 'PPF Account',
+  fd: 'SBI FD',
+  friends: 'Friends Account',
+  flat: 'Flat Shared Account',
+  family: 'Family Account',
+  cashbackPool: 'Cashback Pool',
+  cash: 'Cash',
+  voucher: 'Voucher Account',
+} as const
+
+export interface ExpenseTemplate {
+  category: string
+  subcategory: string
+  min: number
+  max: number
+  account: string
+  freq?: number
+  day?: number
+}
+
+export const EXPENSE_TEMPLATES: readonly ExpenseTemplate[] = [
+  { category: 'Food & Dining', subcategory: 'Office Cafeteria', min: 40, max: 150, account: ACCOUNTS.pluxee, freq: 20 },
+  { category: 'Food & Dining', subcategory: 'Delivery Apps', min: 100, max: 600, account: ACCOUNTS.swiggyCC, freq: 8 },
+  { category: 'Food & Dining', subcategory: 'Dining Out', min: 300, max: 2500, account: ACCOUNTS.swiggyCC, freq: 2 },
+  { category: 'Food & Dining', subcategory: 'Groceries', min: 200, max: 3500, account: ACCOUNTS.gpay, freq: 4 },
+  { category: 'Food & Dining', subcategory: 'Snacks & Beverages', min: 20, max: 200, account: ACCOUNTS.gpay, freq: 6 },
+  { category: 'Transportation', subcategory: 'Daily Commute', min: 25, max: 250, account: ACCOUNTS.gpay, freq: 18 },
+  { category: 'Transportation', subcategory: 'InterCity Travel', min: 500, max: 8000, account: ACCOUNTS.hdfc, freq: 0.3 },
+  { category: 'Housing', subcategory: 'Rent', min: 15000, max: 15000, account: ACCOUNTS.hdfc, day: 5 },
+  { category: 'Housing', subcategory: 'Utilities', min: 800, max: 3000, account: ACCOUNTS.hdfc, freq: 2 },
+  { category: 'Housing', subcategory: 'Household Items', min: 100, max: 2000, account: ACCOUNTS.amazonCC, freq: 1.5 },
+  { category: 'Housing', subcategory: 'Domestic Help', min: 2000, max: 3000, account: ACCOUNTS.cash, day: 1 },
+  { category: 'Housing', subcategory: 'Maintenance & Repairs', min: 500, max: 5000, account: ACCOUNTS.gpay, freq: 0.2 },
+  { category: 'Entertainment & Recreations', subcategory: 'Recharge', min: 179, max: 599, account: ACCOUNTS.gpay, freq: 1.5 },
+  { category: 'Entertainment & Recreations', subcategory: 'OTT Subscriptions', min: 149, max: 649, account: ACCOUNTS.swiggyCC, day: 1 },
+  { category: 'Entertainment & Recreations', subcategory: 'Movies & Events', min: 200, max: 1500, account: ACCOUNTS.axisCC, freq: 0.8 },
+  { category: 'Entertainment & Recreations', subcategory: 'Parties & Treats', min: 300, max: 3000, account: ACCOUNTS.swiggyCC, freq: 0.8 },
+  { category: 'Entertainment & Recreations', subcategory: 'Hobbies', min: 200, max: 2000, account: ACCOUNTS.amazonCC, freq: 0.3 },
+  { category: 'Personal Care', subcategory: 'Clothing', min: 500, max: 5000, account: ACCOUNTS.axisCC, freq: 0.8 },
+  { category: 'Personal Care', subcategory: 'Grooming Products', min: 100, max: 800, account: ACCOUNTS.amazonCC, freq: 0.5 },
+  { category: 'Personal Care', subcategory: 'Haircut & Grooming', min: 200, max: 600, account: ACCOUNTS.gpay, freq: 0.5 },
+  { category: 'Education', subcategory: 'Books & Supplies', min: 200, max: 2000, account: ACCOUNTS.amazonCC, freq: 0.4 },
+  { category: 'Education', subcategory: 'Courses & Workshops', min: 500, max: 10000, account: ACCOUNTS.hdfc, freq: 0.1 },
+  { category: 'Education', subcategory: 'Exam Fees', min: 500, max: 5000, account: ACCOUNTS.hdfc, freq: 0.08 },
+  { category: 'Healthcare', subcategory: 'Medicines & Supplements', min: 100, max: 2000, account: ACCOUNTS.amazonCC, freq: 0.3 },
+  { category: 'Healthcare', subcategory: 'Doctor Visits', min: 500, max: 3000, account: ACCOUNTS.gpay, freq: 0.15 },
+  { category: 'Gadgets & Accessories', subcategory: 'Devices', min: 2000, max: 50000, account: ACCOUNTS.amazonCC, freq: 0.06 },
+  { category: 'Gadgets & Accessories', subcategory: 'Accessories', min: 300, max: 5000, account: ACCOUNTS.axisCC, freq: 0.15 },
+  { category: 'Family', subcategory: 'Regular Support', min: 5000, max: 15000, account: ACCOUNTS.hdfc, freq: 0.5 },
+  { category: 'Family', subcategory: 'Extra Expenses', min: 1000, max: 30000, account: ACCOUNTS.hdfc, freq: 0.3 },
+  { category: 'Charity', subcategory: 'Donations', min: 100, max: 2000, account: ACCOUNTS.gpay, freq: 0.3 },
+  { category: 'Charity', subcategory: 'Religious Offerings', min: 50, max: 500, account: ACCOUNTS.cash, freq: 0.2 },
+  { category: 'Miscellaneous', subcategory: 'Unknown', min: 50, max: 2000, account: ACCOUNTS.sbi, freq: 2 },
+  { category: 'Miscellaneous', subcategory: 'Software Subscriptions', min: 100, max: 1500, account: ACCOUNTS.swiggyCC, freq: 0.3 },
+  { category: 'Miscellaneous', subcategory: 'Bank Fees', min: 50, max: 500, account: ACCOUNTS.sbi, freq: 0.2 },
+  { category: 'Miscellaneous', subcategory: 'Gifts', min: 500, max: 5000, account: ACCOUNTS.amazonCC, freq: 0.3 },
+  { category: 'Investment Charges & Loss', subcategory: 'Brokerage Fees', min: 20, max: 500, account: ACCOUNTS.growStocks, freq: 0.3 },
+] as const
+
+export const NOTES_MAP: Record<string, string[]> = {
+  'Office Cafeteria': ['Sambhar Rice', 'Egg Biryani', 'Lunch', 'Papaya Bowl', 'Thali', 'Curd Rice', 'Dosa', 'Idli Plate', 'Poha', 'Coffee'],
+  'Delivery Apps': ['Mattar Paneer Meal', 'Burger Combo', 'Pizzas + Pepsi', 'Chicken Biryani', 'Pasta Bowl', 'Wraps Combo', 'South Indian Thali', 'Chinese Combo'],
+  'Dining Out': ['Dinner at Barbeque Nation', 'Lunch - Saravana Bhavan', 'Pizza Hut', 'Cafe Coffee Day', 'Team Lunch', 'Birthday Dinner', 'Weekend Brunch'],
+  Groceries: ['BigBasket Order', 'DMart', 'Zepto Order', 'Blinkit Groceries', 'Swiggy Instamart', 'Reliance Fresh'],
+  'Snacks & Beverages': ['Coconut Water', 'Juice - Mango', 'Milk Shake - Banana', 'Chai', 'Cold Coffee', 'Ice Cream', 'Street Food'],
+  'Daily Commute': ['Uber Bike Flat to Office', 'Uber Bike Office to Flat', 'Uber Auto Office to Flat', 'Rapido Bike', 'Metro Recharge'],
+  'InterCity Travel': ['Train Ticket - Weekend Trip', 'Flight - Home Visit', 'Bus - Weekend Getaway', 'Redbus Booking'],
+  Rent: ['Monthly Rent'],
+  Utilities: ['Electricity Bill', 'WiFi Bill', 'Water Bill', 'Gas Cylinder', 'Cooking Gas'],
+  'Household Items': ['Water Bottle', 'Surf Excel', 'Cleaning Supplies', 'Kitchen Items', 'Broom & Mop'],
+  'Domestic Help': ['Maid Monthly', 'Cook Monthly'],
+  Recharge: ['Jio Recharge - Self', 'Jio Data Add-On', 'Airtel Recharge - Self'],
+  'OTT Subscriptions': ['Netflix Subscription', 'Amazon Prime', 'Disney+ Hotstar', 'Spotify Premium', 'YouTube Premium'],
+  'Movies & Events': ['PVR Cinemas', 'BookMyShow', 'Cricket Turf', 'Concert Tickets'],
+  'Parties & Treats': ['Birthday Party', 'Team Dinner', 'Housewarming Treat', 'Farewell Party'],
+  Clothing: ['Myntra Order', 'Flipkart Fashion', 'Amazon Fashion', 'Zara', 'H&M'],
+  'Grooming Products': ['Shampoo', 'Face Cream', 'Facewash', 'Body Wash', 'Sunscreen'],
+  'Haircut & Grooming': ['Haircut & Face Massage', 'Haircut & Trim', 'Beard Trim'],
+  'Books & Supplies': ['Amazon Books', 'Kindle Purchase', 'Tech Book', 'Stationery'],
+  'Courses & Workshops': ['Udemy Course', 'AWS Certification Prep', 'Online Workshop'],
+  'Medicines & Supplements': ['Apollo Pharmacy', '1mg Order', 'Protein Powder', 'Multivitamins'],
+  'Doctor Visits': ['General Checkup', 'Lab Tests', 'Dental Checkup', 'Eye Checkup'],
+  Devices: ['Laptop Upgrade', 'Wireless Earbuds', 'Smart Watch', 'Keyboard'],
+  Accessories: ['Phone Case', 'USB Hub', 'Charger', 'Screen Protector', 'Mouse Pad'],
+  'Regular Support': ['Monthly Family Support', 'Parents - Medical', 'Parents - Groceries'],
+  'Extra Expenses': ['Festival Gift - Family', 'Emergency - Family', 'Home Repair - Parents'],
+  Donations: ['NGO Donation', 'Charity Event', 'Education Fund'],
+  'Religious Offerings': ['Temple Donation', 'Festival Offering'],
+  Unknown: ['Miscellaneous', 'Cash Withdrawal', 'ATM', 'UPI Payment'],
+  Gifts: ['Birthday Gift', 'Wedding Gift', 'Festival Gift', 'Housewarming Gift'],
+  'Brokerage Fees': ['Groww Brokerage', 'DP Charges', 'STT Charges'],
+}
