+++ frontend/src/pages/mutual-fund-projection/types.ts
@@ -1,0 +1,17 @@
+export interface ChartDataPoint {
+  month: string
+  invested: number
+  value: number
+  isHistorical: boolean
+}
+
+export interface MutualFundAccount {
+  name: string
+  balance: number
+}
+
+export interface SIPTransfer {
+  date: string
+  amount: number
+  note?: string | null
+}
